Rearrange mappings to avoid ambiguity.
The default mappings are ambiguous, relying on either a timeout or an
additional keystroke to trigger.  This is most notable for people who,
like me, set notimeout and nottimeout.

Making the append action explicit and consistently putting the action
last makes it clear when the mapped sequence is done, and calculations
complete immediately.